Refactor block_layer

+++ src/db.rs
@@ -43,7 +43,7 @@ impl Db {
         }
 
         let file_layer: FileLayer = FileLayer::new(db_file, Db::PAGE_SIZE);
-        let block_layer: BlockLayer = BlockLayer::new(file_layer, Db::PAGE_SIZE);
+        let block_layer: BlockLayer = BlockLayer::new(file_layer, Db::PAGE_SIZE as usize);
         let page_cache: PageCache = PageCache::new(block_layer, Self::PAGE_SIZE);
 
         let mut db = Db {
